batch init
batch init

+++ src/main/java/com/hejula/server/batch/hejulaserverbatch/entity/Accommodation.java
@@ -5,9 +5,11 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonManagedReference;
 import lombok.*;
 import org.hibernate.annotations.ColumnDefault;
+import org.hibernate.annotations.DynamicUpdate;
 
 import javax.persistence.*;
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
@@ -81,7 +83,7 @@ private long guSeq;
   @Column(nullable = false)
   private long yCoordinate;
 
-  private long yearAveragePrice; //연 평균 가격
+  private double yearAveragePrice; //연 평균 가격
 
   @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-mm-dd HH:mm:ss", timezone = "Asia/Seoul")
   private Date registDate;
